atomicdir: Split testing utils into a helper pkg
