+++ pkg/operator/staticpod/internal/atomicdir/sync_linux_test.go
@@ -8,6 +8,9 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
+
+	adtesting "github.com/openshift/library-go/pkg/operator/staticpod/internal/atomicdir/testing"
+	"github.com/openshift/library-go/pkg/operator/staticpod/internal/atomicdir/types"
 )
 
 func TestSync(t *testing.T) {
@@ -22,11 +25,11 @@ func TestSync(t *testing.T) {
 		newFS func() *fileSystem
 		// existingFiles is used to populate the target directory state before testing.
 		// An empty map will cause the directory to be created, a nil map will cause no directory to be created.
-		existingFiles map[string]File
+		existingFiles map[string]types.File
 		// filesToSync will be synchronized into the target directory.
-		filesToSync map[string]File
+		filesToSync map[string]types.File
 		// expectedFiles contains the files that are expected to be in the target directory after sync is called.
-		expectedFiles map[string]File
+		expectedFiles map[string]types.File
 		// expectSyncError check the return value from Sync.
 		expectSyncError bool
 		// expectLingeringStagingDirectory can be set to true to expect the temporary directory not to be removed.
@@ -37,15 +40,15 @@ func TestSync(t *testing.T) {
 		return testCase{
 			name:  name,
 			newFS: newFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -58,11 +61,11 @@ func TestSync(t *testing.T) {
 			name:          "target directory does not exist",
 			newFS:         newRealFS,
 			existingFiles: nil,
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -70,12 +73,12 @@ func TestSync(t *testing.T) {
 		{
 			name:          "target directory is empty",
 			newFS:         newRealFS,
-			existingFiles: map[string]File{},
-			filesToSync: map[string]File{
+			existingFiles: map[string]types.File{},
+			filesToSync: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -83,15 +86,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "target directory already synchronized",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -99,15 +102,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "change file contents preserving the filenames",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"tls.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
@@ -115,15 +118,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "change filenames preserving the file contents",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"api.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"api.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -131,15 +134,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "change filenames and file contents",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
@@ -147,15 +150,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "replace a single file content",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("3"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("3"), Perm: 0600},
 			},
@@ -163,15 +166,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "replace a single file",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"3.txt": {Content: []byte("3"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"3.txt": {Content: []byte("3"), Perm: 0600},
 			},
@@ -179,15 +182,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "rename a single file",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"3.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"3.txt": {Content: []byte("2"), Perm: 0600},
 			},
@@ -195,17 +198,17 @@ func TestSync(t *testing.T) {
 		{
 			name:  "add new files",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"tls.crt":         {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key":         {Content: []byte("TLS key"), Perm: 0600},
 				"another_tls.crt": {Content: []byte("another TLS cert"), Perm: 0600},
 				"another_tls.key": {Content: []byte("another TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt":         {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key":         {Content: []byte("TLS key"), Perm: 0600},
 				"another_tls.crt": {Content: []byte("another TLS cert"), Perm: 0600},
@@ -215,26 +218,26 @@ func TestSync(t *testing.T) {
 		{
 			name:  "delete a single file",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 			},
 		},
 		{
 			name:  "delete all files",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"1.txt": {Content: []byte("1"), Perm: 0600},
 				"2.txt": {Content: []byte("2"), Perm: 0600},
 			},
-			filesToSync:   map[string]File{},
-			expectedFiles: map[string]File{},
+			filesToSync:   map[string]types.File{},
+			expectedFiles: map[string]types.File{},
 		},
 		errorTestCase("directory unchanged on failed to create object directory", func() *fileSystem {
 			fs := newRealFS()
@@ -286,15 +289,15 @@ func TestSync(t *testing.T) {
 				}
 				return fs
 			},
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key": {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
@@ -304,17 +307,17 @@ func TestSync(t *testing.T) {
 		{
 			name:  "invalid filename specified (relative path)",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				// This fails even though the actual resolved path is just "api.crt".
 				// We simply do not handle paths in any way, we expect filenames.
 				"home/../api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key":         {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -323,15 +326,15 @@ func TestSync(t *testing.T) {
 		{
 			name:  "invalid filename specified (absolute path)",
 			newFS: newRealFS,
-			existingFiles: map[string]File{
+			existingFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
-			filesToSync: map[string]File{
+			filesToSync: map[string]types.File{
 				"/api.crt": {Content: []byte("rotated TLS cert"), Perm: 0600},
 				"api.key":  {Content: []byte("rotated TLS key"), Perm: 0600},
 			},
-			expectedFiles: map[string]File{
+			expectedFiles: map[string]types.File{
 				"tls.crt": {Content: []byte("TLS cert"), Perm: 0600},
 				"tls.key": {Content: []byte("TLS key"), Perm: 0600},
 			},
@@ -341,10 +344,12 @@ func TestSync(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
 			// Write the current directory contents.
 			contentDir := filepath.Join(t.TempDir(), "secrets", "tls-cert")
 			if tc.existingFiles != nil {
-				directoryState(tc.existingFiles).Write(t, contentDir, 0755)
+				adtesting.DirectoryState(tc.existingFiles).Write(t, contentDir, 0755)
 			}
 
 			// Replace with the object data.
@@ -352,7 +357,7 @@ func TestSync(t *testing.T) {
 			err := sync(tc.newFS(), contentDir, 0755, stagingDir, tc.filesToSync)
 
 			// Check the resulting state.
-			directoryState(tc.expectedFiles).CheckDirectoryMatches(t, contentDir, 0755)
+			adtesting.DirectoryState(tc.expectedFiles).CheckDirectoryMatches(t, contentDir, 0755)
 
 			if (err != nil) != tc.expectSyncError {
 				t.Errorf("Expected error from sync = %v, got %v", tc.expectSyncError, err)
